extract reward-info crate
